refactor(app): 미디어 뷰어를 조합 패턴으로 분류한다
