+++ apps/app/src/stories/patterns/PostMediaViewer.stories.tsx
@@ -5,10 +5,10 @@ import { expect, fn, screen, userEvent, waitFor, within } from 'storybook/test';
 import { PostActionAuthenticationProvider } from '@/components/post/PostActionAuthentication';
 import { PostLayout } from '@/components/post/PostLayout';
 import { PostMediaViewerHostProvider } from '@/components/post/PostMediaViewerHost';
-import { PostMediaViewerSurface } from '@/components/post/PostMediaViewerSurface';
 import { PostMediaViewerThread } from '@/components/post/PostMediaViewerThread';
 import { PostReplyCoordinatorProvider } from '@/components/post/PostReplyCoordinator';
 import { ActionMenuPresentationProvider } from '@/components/ui/ActionMenu';
+import { PostMediaViewerSurface } from '@/patterns/post-media-viewer/PostMediaViewerSurface';
 import { SessionProvider } from '@/session/SessionProvider';
 import { useTheme } from '@/theme/ThemeProvider';
 import appleTouchImage from '../../../public/apple-touch-icon.png?url';
